+++ src/commons/queue.rs
@@ -1,12 +1,10 @@
 //! A task queue on top of a key-value store.
 
-use std::{error, fmt};
-use std::borrow::Cow;
-use std::str::FromStr;
+use std::{cmp, error, fmt};
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 use url::Url;
 use crate::commons::storage::{
-    Key, KeyValueError, KeyValueStore, Namespace, Scope, Segment, SegmentBuf,
+    Ident, KeyValueError, KeyValueStore, Transaction
 };
 
 //------------ Configuration -------------------------------------------------
@@ -31,22 +29,22 @@ pub struct Queue {
 }
 
 impl Queue {
-    /// The delay task that have timed out should be reschudeled after.
+    /// The delay task that have timed out should be rescheudled after.
     const RESCHEDULE_AFTER: Duration = Duration::from_secs(15 * 60);
 
     /// The scope for the whole queue.
-    fn lock_scope() -> Scope {
-        Scope::global()
+    const fn lock_scope() -> Option<&'static Ident> {
+        None
     }
 
     /// The scope for pending tasks.
-    fn pending_scope() -> Scope {
-        Scope::from_segment(PendingTask::SEGMENT)
+    const fn pending_scope() -> Option<&'static Ident> {
+        Some(Ident::make("pending"))
     }
 
     /// The scope for running tasks.
-    fn running_scope() -> Scope {
-        Scope::from_segment(RunningTask::SEGMENT)
+    const fn running_scope() -> Option<&'static Ident> {
+        Some(Ident::make("running"))
     }
 }
 
@@ -54,7 +52,7 @@ impl Queue {
     /// Creates a new queue.
     pub fn create(
         storage_uri: &Url,
-        namespace: &Namespace,
+        namespace: &Ident,
     ) -> Result<Self, Error> {
         Ok(Queue {
             store: KeyValueStore::create(storage_uri, namespace)?,
@@ -63,22 +61,22 @@ impl Queue {
 
     /// Returns the number of pending tasks remaining.
     pub fn pending_tasks_remaining(&self) -> Result<usize, Error> {
-        Ok(self.store.execute(&Self::lock_scope(), |kv| {
-            kv.list_keys(&Self::pending_scope()).map(|list| list.len())
+        Ok(self.store.execute(Self::lock_scope(), |kv| {
+            kv.list_keys(Self::pending_scope()).map(|list| list.len())
         })?)
     }
 
     /// Returns the number of running tasks.
     pub fn running_tasks_remaining(&self) -> Result<usize, Error> {
-        Ok(self.store.execute(&Self::lock_scope(), |kv| {
-            kv.list_keys(&Self::running_scope()).map(|list| list.len())
+        Ok(self.store.execute(Self::lock_scope(), |kv| {
+            kv.list_keys(Self::running_scope()).map(|list| list.len())
         })?)
     }
 
     /// Returns the keys of the currently running tasks
-    pub fn running_tasks_keys(&self) -> Result<Vec<Key>, Error> {
-        Ok(self.store.execute(&Self::lock_scope(), |kv| {
-            kv.list_keys(&Self::running_scope())
+    pub fn running_tasks_keys(&self) -> Result<Vec<Box<Ident>>, Error> {
+        Ok(self.store.execute(Self::lock_scope(), |kv| {
+            kv.list_keys(Self::running_scope())
         })?)
     }
 
@@ -91,109 +89,87 @@ impl Queue {
     /// the `mode`.
     pub fn schedule_task(
         &self,
-        name: SegmentBuf,
-        value: serde_json::Value,
+        name: &Ident,
+        value: &serde_json::Value,
         timestamp_millis: Option<u128>,
         mode: ScheduleMode,
     ) -> Result<(), Error> {
-        self.store.execute(&Self::lock_scope(), |s| {
-            let mut new_task = PendingTask {
-                name: name.as_ref(),
-                timestamp_millis: timestamp_millis.unwrap_or(now()),
-                value: &value,
-            };
-            let new_task_key = Key::from(new_task);
+        self.store.execute(Self::lock_scope(), |store| {
+            let mut timestamp =  timestamp_millis.unwrap_or_else(|| {
+                Self::now()
+            });
+            let pending_opt = self.get_storage_key_and_time(
+                name, store, Self::pending_scope()
+            );
+            let running_opt = self.get_storage_key_and_time(
+                name, store, Self::running_scope()
+            );
 
-            let running_key_opt = s
-                .list_keys(&Self::running_scope())?
-                .into_iter()
-                .filter_map(|k| TaskKey::try_from(&k).ok())
-                .find(|running| running.name.as_ref() == new_task.name)
-                .map(|tk| tk.running_key());
-
-            let pending_key_opt = s
-                .list_keys(&Self::pending_scope())?
-                .into_iter()
-                .filter_map(|k| TaskKey::try_from(&k).ok())
-                .find(|p| p.name.as_ref() == new_task.name)
-                .map(|tk| tk.pending_key());
-
-            match mode {
+            let keep = match mode {
                 ScheduleMode::IfMissing => {
-                    if pending_key_opt.is_some()
-                        || running_key_opt.is_some()
-                    {
-                        // nothing to do, there is something
-                        Ok(())
-                    }
-                    else {
-                        // no pending or running task exists, just add
-                        // the new task
-                        s.store(&new_task_key, new_task.value)
-                    }
+                    pending_opt.is_none() && running_opt.is_none()
                 }
                 ScheduleMode::ReplaceExisting => {
-                    if let Some(pending) = pending_key_opt {
-                        s.delete(&pending)?;
+                    if let Some((pending, _)) = pending_opt {
+                        store.delete(Self::pending_scope(), &pending)?;
                     }
-                    s.store(&new_task_key, new_task.value)
+                    true
                 }
                 ScheduleMode::ReplaceExistingSoonest => {
-                    if let Some(pending) = pending_key_opt {
-                        if let Ok(tk) = TaskKey::try_from(&pending) {
-                            new_task.timestamp_millis =
-                                new_task.timestamp_millis.min(
-                                    tk.timestamp_millis
-                                );
-                        }
-                        s.delete(&pending)?;
+                    if let Some((pending, ts)) = pending_opt {
+                        timestamp = cmp::min(
+                            timestamp, ts
+                        );
+                        store.delete(Self::pending_scope(), &pending)?;
                     }
-
-                    let new_task_key = Key::from(new_task);
-                    s.store(&new_task_key, new_task.value)
+                    true
                 }
                 ScheduleMode::FinishOrReplaceExisting => {
-                    if let Some(running) = running_key_opt {
-                        s.delete(&running)?;
+                    if let Some((running, _)) = running_opt {
+                        store.delete(Self::running_scope(), &running)?;
                     }
-                    if let Some(pending) = pending_key_opt {
-                        s.delete(&pending)?;
+                    if let Some((pending, _)) = pending_opt {
+                        store.delete(Self::pending_scope(), &pending)?;
                     }
-                    s.store(&new_task_key, new_task.value)
+                    true
                 }
                 ScheduleMode::FinishOrReplaceExistingSoonest => {
-                    if let Some(running) = running_key_opt {
-                        s.delete(&running)?;
+                    if let Some((running, _)) = running_opt {
+                        store.delete(Self::running_scope(), &running)?;
                     }
-
-                    if let Some(pending) = pending_key_opt {
-                        if let Ok(tk) = TaskKey::try_from(&pending) {
-                            new_task.timestamp_millis =
-                                new_task.timestamp_millis.min(
-                                    tk.timestamp_millis
-                                );
-                        }
-                        s.delete(&pending)?;
+                    if let Some((pending, ts)) = pending_opt {
+                        timestamp = cmp::min(
+                            timestamp, ts
+                        );
+                        store.delete(Self::pending_scope(), &pending)?;
                     }
-
-                    let new_task_key = Key::from(new_task);
-                    s.store(&new_task_key, new_task.value)
+                    true
                 }
+            };
+            if keep {
+                store.store(
+                    Self::pending_scope(),
+                    &Self::task_storage_key(name, Some(timestamp)),
+                    value
+                )?;
             }
+            Ok(())
         })?;
         Ok(())
     }
 
     /// Returns the scheduled timestamp in ms for the named task, if any.
     pub fn pending_task_scheduled(
-        &self, name: &Segment
+        &self, name: &Ident
     ) -> Result<Option<u128>, Error> {
-        Ok(self.store.execute(&Self::lock_scope(), |kv| {
-            kv.list_keys(&Self::pending_scope()).map(|keys| {
-                keys.into_iter()
-                    .filter_map(|k| TaskKey::try_from(&k).ok())
-                    .find(|p| p.name.as_ref() == name)
-                    .map(|p| p.timestamp_millis)
+        Ok(self.store.execute(Self::lock_scope(), |store| {
+            store.list_keys(
+                Self::pending_scope()
+            ).map(|keys| {
+                keys.into_iter().find_map(|key| {
+                    let (ts, key_name) = Self::split_storage_key(&key)?;
+                    (key_name == name).then_some(ts)
+                })
             })
         })?)
     }
@@ -202,15 +178,17 @@ impl Queue {
     ///
     /// Fails if the task is not running.
     pub fn finish_running_task(
-        &self, running_key: &Key
+        &self, storage_key: &Ident
     ) -> Result<(), Error> {
-        self.store.execute(&Self::lock_scope(), |kv| {
-            if kv.has(running_key)? {
-                kv.delete(running_key)?;
+        self.store.execute(Self::lock_scope(), |store| {
+            // XXX This should be done in a single step.
+            if store.has(Self::running_scope(), storage_key)? {
+                store.delete(Self::running_scope(), storage_key)?;
                 Ok(Ok(()))
-            } else {
+            }
+            else {
                 Ok(Err(Error::other(format!(
-                    "Cannot finish task {running_key}. It is not running."
+                    "Cannot finish task {storage_key}. It is not running."
                 ))))
             }
         })?
@@ -220,224 +198,164 @@ impl Queue {
     ///
     /// Fails if the task is not running.
     pub fn reschedule_running_task(
-        &self, running: &Key, timestamp_millis: Option<u128>
+        &self, storage_key: &Ident, timestamp_millis: Option<u128>
     ) -> Result<(), Error> {
-        let pending_key = {
-            let mut task_key = TaskKey::try_from(running)?;
-            task_key.timestamp_millis = timestamp_millis.unwrap_or_else(now);
-
-            task_key.pending_key()
+        let Some((_, name)) = Self::split_storage_key(storage_key) else {
+            return Err(Error::other(format!(
+                "Cannot reschedule task {storage_key}: invalid storage key."
+            )))
         };
-
-        Ok(self.store.execute(&Self::lock_scope(), |kv| {
-            kv.move_value(running, &pending_key)
+        let new_key = Self::task_storage_key(name, timestamp_millis);
+        Ok(self.store.execute(Self::lock_scope(), |store| {
+            store.move_value(
+                Self::running_scope(), storage_key,
+                Self::pending_scope(), &new_key
+            )
         })?)
     }
 
     /// Claims the next scheduled pending task, if any.
     pub fn claim_scheduled_pending_task(
         &self
-    ) -> Result<Option<RunningTask>, Error> {
-        Ok(self.store.execute(&Self::lock_scope(), |kv| {
-            let tasks_before = now();
+    ) -> Result<Option<(Box<Ident>, serde_json::Value)>, Error> {
+        self.store.execute(Self::lock_scope(), |store| {
+            let now = Self::now();
 
-            if let Some(pending) = kv
-                .list_keys(&Self::pending_scope())?
-                .into_iter()
-                .filter_map(|k| TaskKey::try_from(&k).ok())
-                .filter(|tk| tk.timestamp_millis <= tasks_before)
-                .min_by_key(|tk| tk.timestamp_millis)
-            {
-                let pending_key = pending.pending_key();
-
-                if let Some(value) = kv.get(&pending_key)? {
-                    let mut running_task = RunningTask {
-                        name: pending.name.into_owned(),
-                        timestamp_millis: tasks_before,
-                        value,
-                    };
-                    let mut running_key = Key::from(&running_task);
-
-                    if kv.has(&running_key)? {
-                        // It's not pretty to sleep blocking, even if it's
-                        // for 1 ms, but if we don't then get a name collision
-                        // with an existing running task.
-                        std::thread::sleep(Duration::from_millis(1));
-                        running_task.timestamp_millis = now();
-                        running_key = Key::from(&running_task);
+            let Some((_, key)) = store.list_keys(
+                Self::pending_scope()
+            )?.into_iter().fold(None, |acc, key| {
+                let Some((ts, _)) = Self::split_storage_key(&key) else {
+                    return acc
+                };
+                if ts > now {
+                    return acc
+                }
+                if let Some((acc_ts, _)) = acc {
+                    if acc_ts < ts {
+                        return acc
                     }
-
-                    kv.move_value(&pending_key, &running_key)?;
-
-                    Ok(Some(running_task))
                 }
-                else {
-                    Ok(None)
+
+                Some((ts, key))
+            }) else {
+                return Ok(Ok(None))
+            };
+
+            let Some((_, name)) = Self::split_storage_key(&key) else {
+                // We already did this just now, so this is impossible.
+                return Ok(Err(Error::other(format!(
+                    "Cannot load task: storage key '{key}' is suddenly \
+                     invalid. This is a bug."
+                ))));
+            };
+
+            if let Some(value) = store.get(Self::pending_scope(), &key)? {
+                let mut new_key = Self::task_storage_key(name, Some(now));
+
+                if store.has(Self::running_scope(), &new_key)? {
+                    // It's not pretty to sleep blocking, even if it's
+                    // for 1 ms, but if we don't then get a name collision
+                    // with an existing running task.
+                    std::thread::sleep(Duration::from_millis(1));
+                    new_key = Self::task_storage_key(name, None);
                 }
+
+                store.move_value(
+                    Self::pending_scope(), &key,
+                    Self::running_scope(), &new_key
+                )?;
+                
+                Ok(Ok(Some((new_key, value))))
             }
             else {
-                Ok(None)
+                Ok(Ok(None))
             }
-        })?)
+        })?
     }
 
     /// Reschedules running tasks that have timed out.
     pub fn reschedule_long_running_tasks(
-        &self, reschedule_after: Option<&Duration>
+        &self, reschedule_after: Option<Duration>
     ) -> Result<(), Error> {
-        let now = now();
-
         let reschedule_after = reschedule_after.unwrap_or(
-            &Self::RESCHEDULE_AFTER
+            Self::RESCHEDULE_AFTER
         );
-        let reschedule_timeout = now - reschedule_after.as_millis();
+        let reschedule_timeout = Self::now().saturating_sub(
+            reschedule_after.as_millis()
+        );
 
-        Ok(self.store.execute(
-            &Self::lock_scope(),
-            |s| {
-                s.list_keys(&Self::running_scope())?
-                    .into_iter()
-                    .filter_map(|k| {
-                        let task = TaskKey::try_from(&k).ok()?;
-                        if task.timestamp_millis <= reschedule_timeout {
-                            Some(task)
-                        }
-                        else {
-                            None
-                        }
-                    })
-                    .for_each(|tk| {
-                        let running_key = tk.running_key();
-
-                        let pending_key = TaskKey {
-                            name: Cow::Borrowed(&tk.name),
-                            timestamp_millis: now,
-                        }
-                        .pending_key();
-
-                        let _ = s.move_value(&running_key, &pending_key);
-                    });
-
-                Ok(())
-            },
-        )?)
-    }
-}
-
-
-//------------ TaskKey -------------------------------------------------------
-
-/// The key for a task.
-struct TaskKey<'a> {
-    pub name: Cow<'a, Segment>,
-    pub timestamp_millis: u128,
-}
-
-impl TaskKey<'_> {
-    /// Returns the name portion of the key.
-    fn key_name(&self) -> SegmentBuf {
-        SegmentBuf::from_str(&format!(
-            concat!("{}", separator!(), "{}"),
-            self.timestamp_millis, self.name
-        ))
-        .unwrap()
+        Ok(self.store.execute(Self::lock_scope(), |store| {
+            for key in store.list_keys(Self::running_scope())? {
+                let Some((ts, name)) = Self::split_storage_key(&key) else {
+                    continue
+                };
+                if ts <= reschedule_timeout {
+                    let new_key = Self::task_storage_key(name, None);
+                    let _ = store.move_value(
+                        Self::running_scope(), &key,
+                        Self::pending_scope(), &new_key
+                    );
+                }
+            }
+            Ok(())
+        })?)
     }
 
-    fn running_key(&self) -> Key {
-        Key::new_scoped(Queue::running_scope(), self.key_name())
+
+
+    fn now() -> u128 {
+        SystemTime::now().duration_since(
+            UNIX_EPOCH
+        ).map(|d| d.as_millis()).unwrap_or(0)
     }
 
-    fn pending_key(&self) -> Key {
-        Key::new_scoped(Queue::pending_scope(), self.key_name())
+    fn task_storage_key(
+        name: &Ident, timestamp_millis: Option<u128>
+    ) -> Box<Ident> {
+        let timestamp_millis = timestamp_millis.unwrap_or_else(Self::now);
+
+        // Safety: `name` is an ident and a formatted timestamp is only
+        //         digits.
+        unsafe {
+            Ident::boxed_from_string_unchecked(
+                format!(
+                    concat!("{}", separator!(), "{}"),
+                    timestamp_millis, name
+                )
+            )
+        }
     }
-}
 
-impl TryFrom<&Key> for TaskKey<'_> {
-    type Error = InvalidKey;
+    fn split_storage_key(key: &Ident) -> Option<(u128, &Ident)> {
+        let (ts, name) = key.as_str().split_once(separator!())?;
+        if name.is_empty() {
+            return None
+        }
+        let ts = ts.parse().ok()?;
 
-    fn try_from(key: &Key) -> Result<Self, Self::Error> {
-        let (ts, name) = key
-            .name()
-            .as_str()
-            .split_once(separator!())
-            .ok_or(InvalidKey(()))?;
-        Ok(TaskKey {
-            name: Cow::Owned(
-                Segment::parse(name).map_err(|_| InvalidKey(()))?.into()
-            ),
-            timestamp_millis: ts.parse().map_err(|_| InvalidKey(()))?,
+        // Safety: `name` is not empty and came from an Ident so characters
+        //         are fine.
+        //
+        // XXX We should probably move all of this into
+        //     crate::commons::storage::ident to concentrate the unsafe stuff
+        //     there.
+        let name = unsafe { Ident::from_bytes_unchecked(name.as_bytes()) };
+
+        Some((ts, name))
+    }
+
+    fn get_storage_key_and_time(
+        &self, name: &Ident, store: &mut Transaction, scope: Option<&Ident>
+    ) -> Option<(Box<Ident>, u128)> {
+        store.list_keys(scope).ok()?.into_iter().find_map(|key| {
+            let (ts, key_name) = Self::split_storage_key(&key)?;
+            (key_name == name).then_some((key, ts))
         })
     }
 }
 
-impl From<PendingTask<'_>> for Key {
-    fn from(p: PendingTask<'_>) -> Self {
-        let mut key = Key::from_str(&p.to_string()).unwrap();
-        key.add_super_scope(PendingTask::SEGMENT);
-        key
-    }
-}
 
-impl<'a> From<&'a RunningTask> for Key {
-    fn from(p: &'a RunningTask) -> Self {
-        let mut key = Key::from_str(&p.to_string()).unwrap();
-        key.add_super_scope(RunningTask::SEGMENT);
-        key
-    }
-}
-
-
-//------------ PendingTask ---------------------------------------------------
-
-#[derive(Clone, Copy, Debug)]
-pub struct PendingTask<'a> {
-    pub name: &'a Segment,
-    pub timestamp_millis: u128,
-    pub value: &'a serde_json::Value,
-}
-
-impl PendingTask<'static> {
-    const SEGMENT: &'static Segment = Segment::make("pending");
-}
-
-
-impl<'a> PartialEq<PendingTask<'a>> for PendingTask<'_> {
-    fn eq(&self, other: &PendingTask<'a>) -> bool {
-        self.name == other.name
-    }
-}
-
-impl fmt::Display for PendingTask<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f,
-            concat!("{}", separator!(), "{}"),
-            self.timestamp_millis, self.name
-        )
-    }
-}
-
-
-//------------ RunningTask ---------------------------------------------------
-
-#[derive(Clone, Debug)]
-pub struct RunningTask {
-    pub name: SegmentBuf,
-    pub timestamp_millis: u128,
-    pub value: serde_json::Value,
-}
-
-impl RunningTask {
-    const SEGMENT: &'static Segment = Segment::make("running");
-}
-
-impl fmt::Display for RunningTask {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f,
-            concat!("{}", separator!(), "{}"),
-            self.timestamp_millis, self.name
-        )
-    }
-}
+//------------ ScheduleMode --------------------------------------------------
 
 /// Defines scheduling behaviour in case a task by the same name already exists.
 #[derive(Clone, Copy, Debug)]
@@ -467,16 +385,6 @@ pub enum ScheduleMode {
     /// Keep existing pending or running task and in that case do not
     /// add the new task. Otherwise just add the new task.
     IfMissing,
-}
-
-
-//------------ Helpers -------------------------------------------------------
-
-fn now() -> u128 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .expect("time-travel is not supported")
-        .as_millis()
 }
 
 
@@ -550,18 +458,24 @@ mod tests {
     use super::*;
 
     fn queue_store(ns: &str) -> Queue {
-        let storage_url = Url::parse("memory://").unwrap();
-        Queue::create(&storage_url, Namespace::parse(ns).unwrap()).unwrap()
+        Queue::create(
+            &Url::parse("memory://").unwrap(),
+            Ident::from_str(ns).unwrap()
+        ).unwrap()
     }
 
     #[test]
     fn key() {
         assert_eq!(
-            TaskKey {
-                name: Cow::Borrowed(Segment::parse("foo").unwrap()),
-                timestamp_millis: 12,
-            }.key_name(),
-            SegmentBuf::from_str("12-foo").unwrap()
+            Queue::task_storage_key(
+                const { Ident::make("foo") },
+                Some(12)
+            ),
+            Ident::make("12-foo").into()
+        );
+        assert_eq!(
+            Queue::split_storage_key(const { Ident::make("12-foo") }),
+            Some((12, const { Ident::make("foo") }))
         );
     }
 
@@ -575,52 +489,50 @@ mod tests {
                 let queue = queue_store("queue_thread_workers");
 
                 for i in 1..=10 {
-                    let name = &format!("job-{i}");
-                    let segment = Segment::parse(name).unwrap();
+                    let name = Ident::builder(
+                        const { Ident::make("job-") }
+                    ).push_u64(i).finish();
                     let value = Value::from("value");
 
-                    queue
-                        .schedule_task(
-                            segment.into(),
-                            value,
-                            None,
-                            ScheduleMode::FinishOrReplaceExisting,
-                        )
-                        .unwrap();
+                    queue.schedule_task(
+                        &name,
+                        &value,
+                        None,
+                        ScheduleMode::FinishOrReplaceExisting,
+                    ).unwrap();
                     println!("> Scheduled job {}", &name);
                 }
             });
-
             create.join().unwrap();
-            let keys = queue.store.execute(&Scope::global(), |tran| {
-                tran.list_keys(
-                    &Scope::from_segment(PendingTask::SEGMENT)
-                )
+
+            let keys = queue.store.execute(None, |tran| {
+                tran.list_keys(Queue::pending_scope())
             }).unwrap();
             assert_eq!(keys.len(), 10);
 
-            for _i in 1..=10 {
+            for _ in 1..=10 {
                 s.spawn(move || {
                     let queue = queue_store("queue_thread_workers");
 
                     while queue.pending_tasks_remaining().unwrap() > 0 {
-                        if let Some(running_task) = queue.claim_scheduled_pending_task().unwrap() {
-                            queue
-                                .finish_running_task(&Key::from(&running_task))
-                                .unwrap();
+                        if let Some((task_name, _))
+                            = queue.claim_scheduled_pending_task().unwrap()
+                        {
+                            queue.finish_running_task(
+                                &task_name
+                            ).unwrap();
                         }
 
-                        std::thread::sleep(std::time::Duration::from_millis(5));
+                        std::thread::sleep(
+                            std::time::Duration::from_millis(5)
+                        );
                     }
                 });
             }
         });
 
-        let pending = queue.pending_tasks_remaining().unwrap();
-        assert_eq!(pending, 0);
-
-        let running = queue.running_tasks_remaining().unwrap();
-        assert_eq!(running, 0);
+        assert_eq!(queue.pending_tasks_remaining().unwrap(), 0);
+        assert_eq!(queue.running_tasks_remaining().unwrap(), 0);
     }
 
     #[test]
@@ -628,18 +540,12 @@ mod tests {
         let queue = queue_store("test_reschedule_long_running");
         queue.store.wipe().unwrap();
 
-        let name = "job";
-        let segment = Segment::parse(name).unwrap();
+        let name = const { Ident::make("job") };
         let value = Value::from("value");
 
-        queue
-            .schedule_task(
-                segment.into(),
-                value,
-                None,
-                ScheduleMode::FinishOrReplaceExisting,
-            )
-            .unwrap();
+        queue.schedule_task(
+            name, &value, None, ScheduleMode::FinishOrReplaceExisting,
+        ).unwrap();
 
         assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
 
@@ -649,17 +555,14 @@ mod tests {
         assert_eq!(queue.pending_tasks_remaining().unwrap(), 0);
 
         let job = queue.claim_scheduled_pending_task().unwrap();
-
         assert!(job.is_none());
 
-        queue
-            .reschedule_long_running_tasks(Some(&Duration::from_secs(0)))
-            .unwrap();
+        queue.reschedule_long_running_tasks(
+            Some(Duration::from_secs(0))
+        ).unwrap();
 
-        let existing = queue.pending_task_scheduled(segment).unwrap();
-
-        assert!(existing.is_some());
         assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
+        assert!(queue.pending_task_scheduled(name).unwrap().is_some());
 
         let job = queue.claim_scheduled_pending_task().unwrap();
 
@@ -672,36 +575,27 @@ mod tests {
         let queue = queue_store("test_reschedule_finished_task");
         queue.store.wipe().unwrap();
 
-        let name = "task";
-        let segment = Segment::parse(name).unwrap();
+        let name = const { Ident::make("task") };
         let value = Value::from("value");
 
         // Schedule the task
-        queue
-            .schedule_task(
-                segment.into(),
-                value,
-                None,
-                ScheduleMode::FinishOrReplaceExisting,
-            )
-            .unwrap();
+        queue.schedule_task(
+            name, &value, None, ScheduleMode::FinishOrReplaceExisting,
+        ).unwrap();
         assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
 
         // Get the task
-        let running_task = queue.claim_scheduled_pending_task().unwrap().unwrap();
+        let _ = queue.claim_scheduled_pending_task().unwrap().unwrap();
         assert_eq!(queue.pending_tasks_remaining().unwrap(), 0);
         assert_eq!(queue.running_tasks_remaining().unwrap(), 1);
 
         // Finish the task and reschedule
         // queue.finish_running_task(task, Some(rescheduled)).unwrap();
-        queue
-            .schedule_task(
-                running_task.name,
-                running_task.value,
-                Some(now()),
-                ScheduleMode::FinishOrReplaceExisting,
-            )
-            .unwrap();
+        queue.schedule_task(
+            name, &value,
+            Some(Queue::now()),
+            ScheduleMode::FinishOrReplaceExisting,
+        ).unwrap();
 
         // There should now be a new pending task, and the
         // running task should be removed.
@@ -709,11 +603,9 @@ mod tests {
         assert_eq!(queue.running_tasks_remaining().unwrap(), 0);
 
         // Get and finish the pending task, but do not reschedule it
-        let running_task = queue.claim_scheduled_pending_task().unwrap().unwrap();
+        let (key, _) = queue.claim_scheduled_pending_task().unwrap().unwrap();
         assert_eq!(queue.pending_tasks_remaining().unwrap(), 0);
-        queue
-            .finish_running_task(&Key::from(&running_task))
-            .unwrap();
+        queue.finish_running_task(&key).unwrap();
 
         // There should not be a new pending task
         assert_eq!(queue.pending_tasks_remaining().unwrap(), 0);
@@ -724,120 +616,94 @@ mod tests {
         let queue = queue_store("test_schedule_with_existing_task");
         queue.store.wipe().unwrap();
 
-        let name: SegmentBuf = const { Segment::make("task") }.into();
+        let name = const { Ident::make("task") };
         let value_1 = Value::from("value_1");
         let value_2 = Value::from("value_2");
 
-        let in_a_while = now() + 180;
+        let in_a_while = Queue::now() + 180;
 
         // Schedule a task, and then schedule again replacing the old
         {
-            queue
-                .schedule_task(
-                    name.clone(),
-                    value_1.clone(),
-                    None,
-                    ScheduleMode::FinishOrReplaceExisting,
-                )
-                .unwrap();
+            queue.schedule_task(
+                name, &value_1.clone(),
+                None, ScheduleMode::FinishOrReplaceExisting,
+            ).unwrap();
             assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
 
             // Schedule again, replacing the existing task
-            queue
-                .schedule_task(
-                    name.clone(),
-                    value_2.clone(),
-                    None,
-                    ScheduleMode::FinishOrReplaceExisting,
-                )
-                .unwrap();
+            queue.schedule_task(
+                name, &value_2.clone(), None,
+                ScheduleMode::FinishOrReplaceExisting,
+            ).unwrap();
             assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
 
             // We should have one task and the value should match the new task.
-            let task = queue.claim_scheduled_pending_task().unwrap().unwrap();
-            assert_eq!(task.value, value_2);
+            let (key, task_value)
+                = queue.claim_scheduled_pending_task().unwrap().unwrap();
+            assert_eq!(task_value, value_2);
 
             assert_eq!(queue.running_tasks_remaining().unwrap(), 1);
-            queue.finish_running_task(&Key::from(&task)).unwrap();
+            queue.finish_running_task(&key).unwrap();
         }
 
         // Schedule a task, and then schedule again keeping the old
         {
-            queue
-                .schedule_task(
-                    name.clone(),
-                    value_1.clone(),
-                    None,
-                    ScheduleMode::FinishOrReplaceExisting,
-                )
-                .unwrap();
-            queue
-                .schedule_task(
-                    name.clone(),
-                    value_2.clone(),
-                    Some(in_a_while),
-                    ScheduleMode::IfMissing,
-                )
-                .unwrap();
+            queue.schedule_task(
+                name, &value_1.clone(),
+                None, ScheduleMode::FinishOrReplaceExisting,
+            ).unwrap();
+            queue.schedule_task(
+                name, &value_2.clone(),
+                Some(in_a_while), ScheduleMode::IfMissing,
+            ).unwrap();
 
             // there should be only one task, it should not be rescheduled,
             // so we get get it and its value should match old.
             assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
-            let task = queue.claim_scheduled_pending_task().unwrap().unwrap();
-            assert_eq!(task.value, value_1);
+            let (_, task_value)
+                = queue.claim_scheduled_pending_task().unwrap().unwrap();
+            assert_eq!(task_value, value_1);
         }
 
         // Schedule a task, and then schedule again rescheduling it
         {
-            queue
-                .schedule_task(
-                    name.clone(),
-                    value_1.clone(),
-                    None,
-                    ScheduleMode::FinishOrReplaceExisting,
-                )
-                .unwrap();
+            queue.schedule_task(
+                name, &value_1.clone(),
+                None, ScheduleMode::FinishOrReplaceExisting,
+            ).unwrap();
 
             // we expect one pending task
             assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
 
             // reschedule that task to 3 minutes from now, keeping the
             // soonest value
-            queue
-                .schedule_task(
-                    name.clone(),
-                    value_2.clone(),
-                    Some(in_a_while),
-                    ScheduleMode::FinishOrReplaceExistingSoonest,
-                )
-                .unwrap();
+            queue.schedule_task(
+                name, &value_2.clone(),
+                Some(in_a_while),
+                ScheduleMode::FinishOrReplaceExistingSoonest,
+            ).unwrap();
 
             // we still expect one pending task with the earlier
             // time and the new value.
             assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
-            let task = queue.claim_scheduled_pending_task().unwrap().unwrap();
-            assert_eq!(task.value, value_2);
+            let (_, task_value)
+                = queue.claim_scheduled_pending_task().unwrap().unwrap();
+            assert_eq!(task_value, value_2);
 
             // But if we now schedule a task and then reschedule
             // it to 3 minutes from now NOT using the soonest. Then
             // we should see 1 pending task that we cannot claim
             // because it is not due.
-            queue
-                .schedule_task(
-                    name.clone(),
-                    value_1.clone(),
-                    None,
-                    ScheduleMode::FinishOrReplaceExisting,
-                )
-                .unwrap();
-            queue
-                .schedule_task(
-                    name.clone(),
-                    value_1.clone(),
-                    Some(in_a_while),
-                    ScheduleMode::FinishOrReplaceExisting,
-                )
-                .unwrap();
+            queue.schedule_task(
+                name, &value_1.clone(),
+                None,
+                ScheduleMode::FinishOrReplaceExisting,
+            ).unwrap();
+            queue.schedule_task(
+                name, &value_1.clone(),
+                Some(in_a_while),
+                ScheduleMode::FinishOrReplaceExisting,
+            ).unwrap();
 
             assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
             assert!(queue.claim_scheduled_pending_task().unwrap().is_none());
@@ -846,42 +712,38 @@ mod tests {
         // Schedule a task, claim it, and then finish and schedule a new task
         {
             // schedule a task
-            queue
-                .schedule_task(
-                    name.clone(),
-                    value_1.clone(),
-                    None,
-                    ScheduleMode::FinishOrReplaceExisting,
-                )
-                .unwrap();
+            queue.schedule_task(
+                name, &value_1.clone(),
+                None,
+                ScheduleMode::FinishOrReplaceExisting,
+            ).unwrap();
 
             // there should be 1 pending task, and 0 running
             assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
             assert_eq!(queue.running_tasks_remaining().unwrap(), 0);
 
             // claim the task
-            let task = queue.claim_scheduled_pending_task().unwrap().unwrap();
-            assert_eq!(task.value, value_1);
+            let (_, task_value)
+                = queue.claim_scheduled_pending_task().unwrap().unwrap();
+            assert_eq!(task_value, value_1);
             assert_eq!(queue.pending_tasks_remaining().unwrap(), 0);
             assert_eq!(queue.running_tasks_remaining().unwrap(), 1);
 
             // schedule a new task
-            queue
-                .schedule_task(
-                    name.clone(),
-                    value_2.clone(),
-                    None,
-                    ScheduleMode::FinishOrReplaceExisting,
-                )
-                .unwrap();
+            queue.schedule_task(
+                name, &value_2,
+                None, ScheduleMode::FinishOrReplaceExisting,
+            ).unwrap();
 
-            // the running task should now be finished, and there should be 1 new pending task
+            // the running task should now be finished, and there should be
+            // one new pending task
             assert_eq!(queue.running_tasks_remaining().unwrap(), 0);
             assert_eq!(queue.pending_tasks_remaining().unwrap(), 1);
 
             // claim the task, it should match the new task
-            let task = queue.claim_scheduled_pending_task().unwrap().unwrap();
-            assert_eq!(task.value, value_2);
+            let (_, task_value)
+                = queue.claim_scheduled_pending_task().unwrap().unwrap();
+            assert_eq!(task_value, value_2);
         }
     }
 }