+++ src/server/taproxy.rs
@@ -1230,7 +1230,7 @@ mod tests {
         commons::{
             crypto::KrillSignerBuilder,
             eventsourcing::AggregateStore,
-            storage::Namespace,
+            storage::Ident,
             test,
         },
         config::ConfigDefaults,
@@ -1250,14 +1250,14 @@ mod tests {
             let ta_signer_store: AggregateStore<TrustAnchorSigner> =
                 AggregateStore::create(
                     storage_uri,
-                    Namespace::make("ta_signer"),
+                    const { Ident::make("ta_signer") },
                     false,
                 )
                 .unwrap();
             let ta_proxy_store: AggregateStore<TrustAnchorProxy> =
                 AggregateStore::create(
                     storage_uri,
-                    Namespace::make("ta_proxy"),
+                    const { Ident::make("ta_proxy") },
                     false,
                 )
                 .unwrap();